Short is the right entropy
The question is how to do it for the link dendrogram fast.

+++ crc/serial_data.py
@@ -79,7 +79,7 @@ if __name__ == "__main__":
     H.set_colregion(L)
     # Entropy ----
     HS = Hierarchical_Entropy(H.Z, H.nodes)
-    HS.Z2dict("short")
+    HS.Z2dict("long")
     s, sv, sh = HS.S(HS.tree)
     H.entropy = [s, sv, sh]
     for score in opt_score:
